# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BiddingTrack": {
        "description": "FR-16.1 \u2014 what the eligibility engine turns on.\n\n`UNKNOWN` is a first-class answer, not a failure: NFR-LEGAL-1 requires we say\nso plainly rather than guess (eval C5 forbids confident-wrong here).",
        "enum": [
          "ncb",
          "icb",
          "donor",
          "private",
          "unknown"
        ],
        "title": "BiddingTrack",
        "type": "string"
      },
      "EligibilityVerdict": {
        "description": "FR-16.2 / FR-7.6 \u2014 the chip on every card.\n\n`UNKNOWN` must be said plainly and must route to a facilitator, never guessed\n(NFR-LEGAL-1). Ineligible tenders are down-ranked and LABELED, never silently\nhidden \u2014 a bidder who can't see why is a bidder who doesn't trust us.",
        "enum": [
          "eligible",
          "conditional",
          "ineligible",
          "unknown"
        ],
        "title": "EligibilityVerdict",
        "type": "string"
      },
      "LoginIn": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
      "MatchOut": {
        "description": "Public shape for GET /api/v1/matches (org-scoped, FR-7.1).",
        "properties": {
          "eligibility": {
            "$ref": "#/components/schemas/EligibilityVerdict"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "state": {
            "$ref": "#/components/schemas/MatchState"
          },
          "tender": {
            "$ref": "#/components/schemas/TenderOut"
          },
          "tender_id": {
            "format": "uuid",
            "title": "Tender Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tender_id",
          "score",
          "explanation",
          "state",
          "eligibility",
          "tender"
        ],
        "title": "MatchOut",
        "type": "object"
      },
      "MatchState": {
        "enum": [
          "new",
          "saved",
          "dismissed"
        ],
        "title": "MatchState",
        "type": "string"
      },
      "MatchStateFilter": {
        "description": "What a client may ASK `GET /matches` for (docs/11 MAT-1).\n\n`dismissed` is deliberately absent rather than merely unsupported: FR-7.3\nmakes dismissal terminal for the feed, so an endpoint that could list\ndismissed matches would be the first step toward resurfacing them.",
        "enum": [
          "new",
          "saved"
        ],
        "title": "MatchStateFilter",
        "type": "string"
      },
      "MeOut": {
        "properties": {
          "org": {
            "$ref": "#/components/schemas/OrgOut"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "org"
        ],
        "title": "MeOut",
        "type": "object"
      },
      "OrgOut": {
        "properties": {
          "country": {
            "title": "Country",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "org_type": {
            "$ref": "#/components/schemas/OrgType"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "org_type",
          "country",
          "timezone"
        ],
        "title": "OrgOut",
        "type": "object"
      },
      "OrgType": {
        "description": "FR-1.6 \u2014 the most load-bearing enum in the product.\n\nDrives eligibility logic (M16), currency display, and channel defaults. A\ntender an org's type cannot bid on is down-ranked and labeled, never silently\nhidden (FR-7.6).",
        "enum": [
          "local",
          "diaspora",
          "foreign"
        ],
        "title": "OrgType",
        "type": "string"
      },
      "ProblemDetail": {
        "description": "RFC 7807 problem+json. Every error from this API has this shape. Branch on `type` (a stable URL ending in the error code), never on `title` or on the prose in `detail`.",
        "properties": {
          "detail": {
            "description": "What went wrong this time. Prose; may change.",
            "type": "string"
          },
          "instance": {
            "description": "Request path the failure applies to.",
            "type": "string"
          },
          "status": {
            "description": "Repeats the HTTP status code.",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable summary of the code.",
            "type": "string"
          },
          "type": {
            "description": "`https://adera.bid/errors/<code>` \u2014 the stable, matchable identity.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
      "RegisterIn": {
        "properties": {
          "country": {
            "title": "Country",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "org_name": {
            "title": "Org Name",
            "type": "string"
          },
          "org_type": {
            "$ref": "#/components/schemas/OrgType"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "timezone": {
            "default": "Africa/Addis_Ababa",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "org_name",
          "org_type",
          "country"
        ],
        "title": "RegisterIn",
        "type": "object"
      },
      "RegisterOut": {
        "description": "AUTH-1 response. Carries `org` and `audience_note` (ADR-029) so a\n`local`-type org learns immediately what it can and can't do here rather\nthan discovering an empty feed later \u2014 a silent gap is indistinguishable\nfrom a bug (see `identity.service.require_bidder_audience`).",
        "properties": {
          "audience_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience Note"
          },
          "org": {
            "$ref": "#/components/schemas/OrgOut"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "org"
        ],
        "title": "RegisterOut",
        "type": "object"
      },
...
      "TenderOut": {
        "description": "One tender, public fields only (facts + link-out \u2014 FR-9.1).",
        "properties": {
          "bid_bond_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Bond Currency"
          },
          "bid_bond_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Bond Minor"
          },
          "bidding_track": {
            "$ref": "#/components/schemas/BiddingTrack"
          },
          "buyer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer"
          },
          "closing_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closing At"
          },
          "doc_price_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Price Currency"
          },
          "doc_price_minor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Price Minor"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "opening_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opening At"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "buyer",
          "summary",
          "region",
          "language",
          "url",
          "published_at",
          "closing_at",
          "opening_at",
          "bid_bond_minor",
          "bid_bond_currency",
          "doc_price_minor",
          "doc_price_currency",
          "bidding_track"
        ],
        "title": "TenderOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_verified"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationProblem": {
        "description": "A ProblemDetail with per-field validation detail attached.",
        "properties": {
          "detail": {
            "description": "What went wrong this time. Prose; may change.",
            "type": "string"
          },
          "errors": {
            "description": "Pydantic's per-field errors: `loc`, `msg`, `type`.",
            "items": {
              "type": "object"
            },
            "type": "array"
          },
          "instance": {
            "description": "Request path the failure applies to.",
            "type": "string"
          },
          "status": {
            "description": "Repeats the HTTP status code.",
            "type": "integer"
          },
          "title": {
            "description": "Human-readable summary of the code.",
            "type": "string"
          },
          "type": {
            "description": "`https://adera.bid/errors/<code>` \u2014 the stable, matchable identity.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ValidationProblem",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ADERA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_route_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`unauthenticated` \u2014 no session cookie, or it is expired or revoked"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Login Route",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout_route_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "adera_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Adera Session"
            }
          },
          {
            "in": "cookie",
            "name": "adera_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Adera Csrf"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`unauthenticated` \u2014 no session cookie, or it is expired or revoked"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`csrf_failed` \u2014 missing or mismatched X-CSRF-Token on an unsafe method"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Logout Route",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Session restore. Resolves the org through `current_org`, the same\ndependency every org-scoped route uses, so `/me` and (say) `/matches` can\nnever disagree about which org the caller is acting as. It previously took\nthe FIRST membership row, which for a multi-org user silently named an org\nthat `/matches` would then refuse to serve without `?org_id=`.",
        "operationId": "me_route_api_v1_auth_me_get",
        "parameters": [
          {
            "in": "query",
            "name": "org_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Org Id"
            }
          },
          {
            "in": "cookie",
            "name": "adera_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Adera Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`org_id_required` \u2014 the user belongs to several orgs; repeat the call with ?org_id="
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`unauthenticated` \u2014 no session cookie, or it is expired or revoked"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`forbidden` \u2014 authenticated, but not permitted (e.g. no organization)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`not_found` \u2014 no such resource \u2014 also returned instead of 403 across orgs, deliberately"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Me Route",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_route_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`conflict` \u2014 duplicate email, or an illegal state transition"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "Register Route",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/matches": {
      "get": {
        "operationId": "list_matches_api_v1_matches_get",
        "parameters": [
          {
            "description": "filter to one state; omit for everything not dismissed",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MatchStateFilter"
                },
                {
                  "type": "null"
                }
              ],
              "description": "filter to one state; omit for everything not dismissed",
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "org_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Org Id"
            }
          },
          {
            "in": "cookie",
            "name": "adera_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Adera Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchOut"
                  },
                  "title": "Response List Matches Api V1 Matches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`org_id_required` \u2014 the user belongs to several orgs; repeat the call with ?org_id="
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`unauthenticated` \u2014 no session cookie, or it is expired or revoked"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`forbidden` \u2014 authenticated, but not permitted (e.g. no organization)\n\n`audience_restricted` \u2014 org_type=local is supply-side only (ADR-029); bidder features are for diaspora/foreign orgs. Render a named blocked state, never an empty list"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`not_found` \u2014 no such resource \u2014 also returned instead of 403 across orgs, deliberately"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblem"
                }
              }
            },
            "description": "`validation_error` \u2014 the request did not match the schema; see the `errors` member"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "`rate_limited` \u2014 too many requests in one minute; honour the Retry-After header"
          }
        },
        "summary": "List Matches",
        "tags": [
          "matches"
        ]
      }
    },
...
  }
}